requirements.txt
aiohappyeyeballs==2.4.2 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.10.7 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.11" and python_version < "4"
cachetools==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4"
cloudevents==1.11.0 ; python_version >= "3.11" and python_version < "4"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4" and platform_system == "Windows"
decorator==5.1.1 ; python_version >= "3.11" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.11" and python_version < "4"
flask==3.0.3 ; python_version >= "3.11" and python_version < "4"
frozenlist==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
fsspec==2024.9.0 ; python_version >= "3.11" and python_version < "4.0"
functions-framework==3.8.1 ; python_version >= "3.11" and python_version < "4"
gcsfs==2024.9.0.post1 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.20.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.20.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth-oauthlib==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.35.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-bigquery==3.25.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-storage==2.18.2 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.66.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.66.1 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.11" and python_version < "4" and platform_system != "Windows"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4"
jinja2==3.1.4 ; python_version >= "3.11" and python_version < "4"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4"
multidict==6.1.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.1.1 ; python_version >= "3.11" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.11" and python_version < "4.0"
pyarrow==17.0.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.11" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
watchdog==5.0.2 ; python_version >= "3.11" and python_version < "4"
werkzeug==3.0.4 ; python_version >= "3.11" and python_version < "4"
yarl==1.13.1 ; python_version >= "3.11" and python_version < "4.0"
